% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom.R
\name{dependenciesInstalled}
\alias{dependenciesInstalled}
\alias{installDependencies}
\title{Checks for/install dependencies}
\usage{
dependenciesInstalled()

installDependencies()
}
\value{
\code{TRUE} when all dependencies are fulfilled; otherwise, \code{FALSE}.
}
\description{
\code{dependenciesInstalled()} that all the required system dependency,
PhantomJS, is installed, and \code{installDependencies()} installs it if needed.
For more information about where PhantomJS will be installed
see \code{\link[webdriver:install_phantomjs]{webdriver::install_phantomjs()}}.
}
\keyword{internal}
