% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-diff.R
\name{diffviewer_widget}
\alias{diffviewer_widget}
\title{Creat an htmlwidget that shows differences between files or directories}
\usage{
diffviewer_widget(old, new, width = NULL, height = NULL,
  pattern = NULL, preprocess = NULL)
}
\arguments{
\item{old, new}{Names of the old and new directories to compare.
Alternatively, they can be a character vectors of specific files to
compare.}

\item{width}{Width of the htmlwidget.}

\item{height}{Height of the htmlwidget}

\item{pattern}{A filter to apply to the old and new directories.}

\item{preprocess}{Preliminary processing for content filtering.}
}
\description{
This function can be used for viewing differences between current test
results and the expected results
}
