% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-load.R
\name{loadTest}
\alias{loadTest}
\title{Run load tests for a Shiny application}
\usage{
loadTest(testFile = "./tests/myloadtest.R", numConcurrent = 4,
  numTotal = 8, url = NULL)
}
\arguments{
\item{testFile}{The file containing a test script to run. Test script can be
generated from recordTest(load_mode = TRUE).}

\item{numConcurrent}{Number of concurrent connections to simulate.}

\item{numTotal}{Total number of connections to simulation.}

\item{url}{Web address of the deployed Shiny application.}
}
\description{
Run load tests for a Shiny application
}
\details{
This function simulates load against a deployed application. The
  function creates a cluster of workers using the parallel function
  makePSOCKcluster. The number of works is equal to the number of desired
  concurrent connections. Each worker launches a phantomJS process that calls
  the URL and drives the app through the test. Test should be generated using
  the recordTest function load_mode = TRUE. Timing information is aggregated
  and returned as a data frame. If the number of total tests > number of
  concurrent tests, then as workers finish their tests they will start a new
  test. The phantomJS process is recycled, but a new browser session is
  started for each test.
}

