% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{osName}
\alias{osName}
\title{Get the name of the OS}
\usage{
osName()
}
\description{
Returns the name of the current OS. This can be useful for the \code{suffix} when
running \code{\link[=testApp]{testApp()}}.
}
