% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{preprocessFilter}
\alias{preprocessFilter}
\title{Build a preprocessing filter to avoid false negatives in tests.Compare current and expected snapshots}
\usage{
preprocessFilter(images = TRUE, normalize_data = FALSE,
  ignore_keys = NULL, ignore_text = NULL)
}
\arguments{
\item{images}{Should screenshots and PNG images be compared? It can be useful
to set this to \code{FALSE} when the expected results were taken on a
different platform from the current results.}

\item{normalize_data}{This will pre-process the JSON content to
canonicalize it (alphabetical order), so changes of JSON objects order will
no longer be considered as differences. It can be useful to set
this to \code{TRUE} when the content of snapshot is quite heavy
(which means that the snapshooted page may be loaded hieratically).}

\item{ignore_keys}{This will pre-process the JSON content to remove elements
matching these key patterns.}

\item{ignore_text}{This will pre-process the content to ignore text
matching thess pattern (using gsub to replace it by arbitrary value).}
}
\description{
Build a preprocessing filter to avoid false negatives in tests.Compare current and expected snapshots
}
