% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_port}
\alias{random_port}
\title{Generate a random port number}
\usage{
random_port(min = 3000, max = 9000)
}
\arguments{
\item{min}{lower limit}

\item{max}{upper limit}
}
\value{
integer scalar, generated port number
}
\description{
Ideally, phantomjs should support a random port, assigned by the OS.
But it does not, so we generate one randomly and hope for the best.
}
\keyword{internal}
