% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{record_test}
\alias{record_test}
\title{Launch test event recorder for a Shiny app}
\usage{
record_test(app, save_dir = NULL)
}
\arguments{
\item{app}{A \code{\link{shinyapp}} object, or path to}

\item{save_dir}{A directory to save stuff.}
}
\description{
Launch test event recorder for a Shiny app
}

