% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{recordTest}
\alias{recordTest}
\title{Launch test event recorder for a Shiny app}
\usage{
recordTest(app, save_dir = NULL)
}
\arguments{
\item{app}{A \code{\link{ShinyDriver}} object, or path to}

\item{save_dir}{A directory to save stuff.}
}
\description{
Launch test event recorder for a Shiny app
}
