% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{registerInputProcesser}
\alias{registerInputProcesser}
\title{Register an input processor for the test recorder}
\usage{
registerInputProcesser(inputType, processor)
}
\arguments{
\item{inputType}{The name of an input type, for example, \code{"mypkg:numberinput"}.}

\item{processor}{An input processor function.}
}
\description{
This function registers an input processor which will be used by the test
recorder. The input processor function should take one parameter, \code{value},
and return a string of R code which returns the desired value.
}
