% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{registerInputProcessor}
\alias{registerInputProcessor}
\alias{getInputProcessors}
\title{Register an input processor for the test recorder}
\usage{
registerInputProcessor(inputType, processor)

getInputProcessors()
}
\arguments{
\item{inputType}{The name of an input type, for example,
\code{"mypkg.numberinput"}.}

\item{processor}{An input processor function.}
}
\description{
\code{registerInputProcessor()} registers an input processor which will be used by
the test recorder. The input processor function should take one parameter,
\code{value}, and return a string of R code which returns the desired value.

\code{getInputProcessors()} returns a named list of all registered input processor
functions.
}
\keyword{internal}
