% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-mapping.R
\name{sd_findWidget}
\alias{sd_findWidget}
\title{Try to deduce the shiny input/output element type from its name}
\usage{
sd_findWidget(self, private, name, iotype)
}
\arguments{
\item{self}{me}

\item{private}{private me}

\item{name}{The name of the Shiny input or output to search for.}

\item{iotype}{It is possible that an input has the same name as
an output, and in this case there is no way to get element without
knowing whether it is an input or output element.}
}
\description{
Try to deduce the shiny input/output element type from its name
}
\keyword{internal}
