% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{shinyapp}
\alias{shinyapp}
\title{Class to manage shiny app and phantom.js processes for testing}
\description{
Class to manage shiny app and phantom.js processes for testing
}
\section{Usage}{


TODO
}

\section{Arguments}{


TODO
}

\section{Details}{


TODO
}

