% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest-package.R
\docType{package}
\name{shinytest-package}
\alias{shinytest}
\alias{shinytest-package}
\title{shinytest: Test Shiny Apps}
\description{
For automated testing of Shiny applications, using
    a headless browser, driven through 'WebDriver'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/shinytest}
  \item Report bugs at \url{https://github.com/rstudio/shinytest/issues}
}

}
\author{
\strong{Maintainer}: Winston Chang \email{winston@rstudio.com}

Authors:
\itemize{
  \item Gábor Csárdi \email{gabor@rstudio.com}
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
  \item Mango Solutions [copyright holder, conceptor]
}

}
\keyword{internal}
