% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-app.R
\name{test_app}
\alias{test_app}
\title{Run tests for a Shiny application}
\usage{
test_app(appDir)
}
\arguments{
\item{appDir}{Path to the Shiny application to be tested.}
}
\description{
Run tests for a Shiny application
}

