% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{widget}
\alias{widget}
\title{Class for a Shiny widget}
\description{
Class for a Shiny widget
}
\section{Usage}{

\preformatted{w <- app$find_widget(name,
    iotype = c("auto", "input", "output"))

w$get_name()
w$get_element()
w$get_type()
w$get_iotype()
w$in_input()
w$is_output()

w$get_value()
w$set_value(value)

w$send_keys(keys)

w$list_tabs()
}
}

\section{Arguments}{

\describe{
  \item{app}{A \code{\link{shinyapp}} object.}
  \item{w}{A \code{widget} object.}
  \item{name}{Name of a Shiny widget.}
  \item{iotype}{Character scalar, whether the widget is \sQuote{input}
    or \sQuote{output}. The default \sQuote{auto} value works well,
    provided that widgets have unique names. (Shiny allows an input
    and an output widget with the same name.)}
  \item{value}{Value to set for the widget. Its interpretation depends
    on the type of the widget, see details below.}
  \item{keys}{Keys to send to the widget. See the \code{send_keys}
    method of the \code{\link[webdriver]{element}} class in the
    \code{webdriver} package.}
}
}

\section{Details}{


A \code{widget} object represents a Shiny input or output widget.
\code{app$find_widget} creates a widget object from a
\code{\link{shinyapp}} object.

\code{w$get_name()} returns the name of the widget.

\code{w$get_element()} returns an HTML element. This is an
\code{\link[webdriver]{element}} object from the \code{webdriver}
package.

\code{w$get_type()} returns the type of the widget, possible values
are \code{textInput}, \code{selectInput}, etc.

\code{w$get_iotype()} returns \sQuote{input} or \sQuote{output},
whether the widget is an input or output widget.

\code{w$is_input()} returns \code{TRUE} for input widgets, \code{FALSE}
otherwise.

\code{w$is_output()} returns \code{TRUE} for output widgets, \code{FALSE}
otherwise.

\code{w$get_value()} returns the value of the widget. The exact type
returned depends on the type of the widget. TODO: list widgets and their
return types.

\code{w$set_value()} sets the value of the widget, through the web
browser. Different widget types expect different different \code{value}
arguments. TODO: list widgets and types.

\code{w$send_keys} sends the specified keys to the HTML element of the
widget.

\code{w$list_tabs} lists the tab names of a \code{tabsetPanel} widget.
It fails for other types of widgets.
}
\examples{
{

}
}

