% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{app_start_phantomjs}
\alias{app_start_phantomjs}
\title{Start phantomjs}
\usage{
app_start_phantomjs(self, private, debug_level)
}
\arguments{
\item{self}{me}

\item{private}{dark side of me}

\item{debug_level}{debug level}
}
\description{
It is not possible to start phantomjs on a randomized port currently,
unfortunately.
}
\details{
`processx::process` will automatically kill it, once `app` is
garbage collected.
}
\keyword{internal}

