% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{dependenciesInstalled}
\alias{dependenciesInstalled}
\title{Checks all dependencies are installed}
\usage{
dependenciesInstalled()
}
\value{
\code{TRUE} when all dependencies are fulfilled; otherwise,
  \code{FALSE}.
}
\description{
Checks that all the required system dependencies are installed properly,
returns. If dependencies are missing, consider running
\link{installDependencies}.
}
\seealso{
\code{\link{installDependencies}} to install missing dependencies.
}
