% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installDependencies}
\alias{installDependencies}
\title{Installs missing dependencies}
\usage{
installDependencies()
}
\description{
Installs all the required system depencies to record and run tests. This will
install a headless web browser, PhantomJS.
}
\examples{
\dontrun{

if (!dependenciesInstalled() &&
    identical(menu(c("Yes", "No"), "Install missing dependencies?"), 1L)) {
  installDependencies()
}

}

}
\seealso{
\code{\link{dependenciesInstalled}} to check if dependencies are
  missing. For more information about where PhantomJS will be installed, see
  \code{\link[webdriver]{install_phantomjs}}.
}
