% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{recordTest}
\alias{recordTest}
\title{Launch test event recorder for a Shiny app}
\usage{
recordTest(app = ".", save_dir = NULL, load_mode = FALSE)
}
\arguments{
\item{app}{A \code{\link{ShinyDriver}} object, or path to a Shiny
application.}

\item{save_dir}{A directory to save stuff.}

\item{load_mode}{A boolean that determines whether or not the resulting test
script should be appropriate for load testing.}
}
\description{
Launch test event recorder for a Shiny app
}
