% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder.R
\name{recordTest}
\alias{recordTest}
\title{Launch test event recorder for a Shiny app}
\usage{
recordTest(app = ".", save_dir = NULL, load_mode = FALSE, seed = NULL,
  loadTimeout = 10000, shinyOptions = list())
}
\arguments{
\item{app}{A \code{\link{ShinyDriver}} object, or path to a Shiny
application.}

\item{save_dir}{A directory to save stuff.}

\item{load_mode}{A boolean that determines whether or not the resulting test
script should be appropriate for load testing.}

\item{seed}{A random seed to set before running the app. This seed will also
be used in the test script.}

\item{loadTimeout}{Maximum time to wait for the Shiny application to load, in
milliseconds. If a value is provided, it will be saved in the test script.}

\item{shinyOptions}{A list of options to pass to \code{runApp()}. If a value
is provided, it will be saved in the test script.}
}
\description{
Launch test event recorder for a Shiny app
}
