% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest-package.R
\docType{package}
\name{shinytest-package}
\alias{shinytest}
\alias{shinytest-package}
\title{shinytest: Test Shiny Apps}
\description{
Please see the shinytest to shinytest2 migration guide at https://rstudio.github.io/shinytest2/articles/z-migration.html
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/shinytest}
  \item Report bugs at \url{https://github.com/rstudio/shinytest/issues}
}

}
\author{
\strong{Maintainer}: Winston Chang \email{winston@posit.co}

Authors:
\itemize{
  \item Gábor Csárdi \email{gabor@posit.co}
  \item Hadley Wickham \email{hadley@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Ascent Digital Services [copyright holder, conceptor]
}

}
\keyword{internal}
