% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabset}
\alias{tabset}
\title{Class to manage tabsets of a Shiny app}
\description{
Class to manage tabsets of a Shiny app
}
\section{Usage}{

\preformatted{ts <- app$get_tabset(id)
tslist <- app$get_tabsets()

ts$list_tabs()
ts$get_active_tab()
ts$set_active_tab(id)
}
}

\section{Arguments}{

\describe{
  \item{app}{A \code{\link{shinyapp}} object.}
  \item{ts}{A \code{tabset} object.}
  \item{tslist}{A list of \code{tabset} objects.}
  \item{id}{For \code{app$get_tabset} a string, the id of the
    \code{tabsetPanel} in Shiny. For \code{ts$set_active_tab} the value
    (or title if value is not given) of the \code{tabPanel} in Shiny.}
}
}

\section{Details}{


A \code{tabset} object represents a Shiny \code{tabsetPanel}.

\code{app$get_tabset()} creates a single \code{tabset} object and can
be used for tabset panels that have an id.

\code{app$get_tabsets()} returns all tabset, regardless of whether they
have ids or not.

\code{ts$list_tabs()} lists all tab values (or titles if no values are
given in the Shiny app) of a tabset.

\code{ts$get_active_tab()} returns the value (or title if no value was
given in the Shiny app) of a tabset.

\code{ts$set_active_tab()} sets the active tab to the one specified
by a value (or title if no value was given in the Shiny app).
}

