% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-app.R
\name{testApp}
\alias{testApp}
\title{Run tests for a Shiny application}
\usage{
testApp(appDir, scripts = NULL)
}
\arguments{
\item{appDir}{Path to the Shiny application to be tested.}

\item{scripts}{Test script(s) to run. For example, \code{"mytest"} or
\code{c("mytest", "mytest2")}. If \code{NULL} (the default), all scripts in
the tests/ directory will be run.}
}
\description{
Run tests for a Shiny application
}

