% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-app.R
\name{testApp}
\alias{testApp}
\title{Run tests for a Shiny application}
\usage{
testApp(appDir = ".", files = NULL, quiet = FALSE,
  compareScreenshot = TRUE)
}
\arguments{
\item{appDir}{Path to the Shiny application to be tested.}

\item{files}{Test script(s) to run. The .R extension of the filename is
optional. For example, \code{"mytest"} or \code{c("mytest", "mytest2.R")}.
If \code{NULL} (the default), all scripts in the tests/ directory will be
run.}

\item{quiet}{Should output be suppressed? This is useful for automated
testing.}

\item{compareScreenshot}{Should screenshots be compared? It can be useful to
set this to \code{FALSE} when the expected results were taken on a
different platform from the one currently being used to test the
application.}
}
\description{
Run tests for a Shiny application
}
