% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-app.R
\name{testApp}
\alias{testApp}
\title{Run tests for a Shiny application}
\usage{
testApp(appDir = ".", files = NULL)
}
\arguments{
\item{appDir}{Path to the Shiny application to be tested.}

\item{files}{Test script(s) to run. For example, \code{"mytest.R"} or
\code{c("mytest.R", "mytest2.R")}. If \code{NULL} (the default), all
scripts in the tests/ directory will be run.}
}
\description{
Run tests for a Shiny application
}
