% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-diff.R
\name{textTestDiff}
\alias{textTestDiff}
\title{Get textual diff of test results}
\usage{
textTestDiff(appDir = ".", testnames = NULL, preprocess = NULL,
             images = TRUE, suffix = NULL)
}
\arguments{
\item{appDir}{Directory of the Shiny application that was tested.}

\item{testnames}{A character vector of names of tests to compare. If NULL,
compare all test results for which there are differences.}

\item{preprocess}{Preliminary processing for content filtering.}
\item{images}{Compare screenshot images.}
\item{suffix}{An optional suffix for the expected results directory. For
example, if the suffix is \code{"mac"}, the expected directory would be
\code{mytest-expected-mac}.}
}
\description{
Get textual diff of test results
}
\seealso{
\code{\link{viewTestDiff}} for interactive diff viewer.
}
