% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-diff.R
\name{viewTestDiff}
\alias{viewTestDiff}
\title{Interactive viewer for changes in test results}
\usage{
viewTestDiff(appDir = ".", testname = NULL)
}
\arguments{
\item{appDir}{Directory of the Shiny application that was tested.}

\item{testname}{Name of test to compare.}
}
\description{
Interactive viewer for changes in test results
}
